NTP-640: comparing and contrasting t ps desktop (#192)
* NTP-640: refactor code so class variable names are consistent

* NTP-640: refactor code to move records/classes, enums and extensions in to own files and folders

* NTP-640: refactor code bug fix

* NTP-640: Minor refactor so using enum rather than string
