+++ UI/Enums/TuitionType.cs
@@ -1,0 +1,16 @@
+﻿using System.ComponentModel;
+
+namespace UI.Enums
+{
+    public enum TuitionType
+    {
+        [Description("Any")]
+        Any = 0,
+
+        [Description("Online")]
+        Online = 1,
+
+        [Description("In School")]
+        InSchool,
+    }
+}
