+++ UI/Models/SearchModel.cs
@@ -1,0 +1,34 @@
+﻿using UI.Enums;
+
+namespace UI.Models
+{
+    public record SearchModel
+    {
+        public SearchModel()
+        {
+
+        }
+
+        public SearchModel(SearchModel model) : base()
+        {
+            From = model.From;
+            Name = model.Name;
+            Postcode = model.Postcode;
+            Subjects = model.Subjects;
+            TuitionType = model.TuitionType;
+            KeyStages = model.KeyStages;
+        }
+
+        public ReferrerList From { get; set; }
+
+        public string? Name { get; set; }
+
+        public string? Postcode { get; set; }
+
+        public string[]? Subjects { get; set; }
+
+        public TuitionType? TuitionType { get; set; }
+
+        public KeyStage[]? KeyStages { get; set; }
+    }
+}
