+++ UI/Models/Selectable.cs
@@ -1,0 +1,10 @@
+﻿namespace UI.Models
+{
+    public class Selectable : Selectable<string> { }
+
+    public class Selectable<T>
+    {
+        public T Name { get; set; } = default!;
+        public bool Selected { get; set; }
+    }
+}
