--- UI/Pages/SearchModel.cs
@@ -1,135 +1,0 @@
-﻿using System.ComponentModel;
-using System.Diagnostics.CodeAnalysis;
-using System.Text.RegularExpressions;
-
-namespace UI.Pages;
-
-public record SearchModel
-{
-    public SearchModel()
-    {
-
-    }
-
-    public SearchModel(SearchModel model) : base()
-    {
-        From = model.From;
-        Name = model.Name;
-        Postcode = model.Postcode;
-        Subjects = model.Subjects;
-        TuitionType = model.TuitionType;
-        KeyStages = model.KeyStages;
-    }
-
-    public ReferrerList From { get; set; }
-
-    public string? Name { get; set; }
-
-    public string? Postcode { get; set; }
-
-    public string[]? Subjects { get; set; }
-
-    public TuitionType? TuitionType { get; set; }
-
-    public KeyStage[]? KeyStages { get; set; }
-}
-
-public static class ModelExtensions
-{
-    public static KeyStageSubject[] ParseKeyStageSubjects(this string[] keyStageSubjects)
-        => keyStageSubjects.Select(KeyStageSubject.TryParse)
-        .OfType<KeyStageSubject>()
-        .ToArray();
-}
-
-public record KeyStageSubject(KeyStage KeyStage, string Subject)
-{
-    public static KeyStageSubject Parse(string value)
-    {
-        var re = new Regex(@"(KeyStage[\d])-(.*)").Match(value);
-
-        if (!re.Success)
-            throw new ArgumentException("Subject must be of the form KS1-English");
-
-        if (!Enum.TryParse<KeyStage>(re.Groups[1].Value, out var ks))
-            throw new ArgumentException($"'{re.Groups[0].Value}' is not a valid key stage");
-
-        return new KeyStageSubject(ks, re.Groups[2].Value);
-    }
-
-    public static KeyStageSubject? TryParse(string value) =>
-        TryParse(value, out var parsed) ? parsed : null;
-
-    public static bool TryParse(string value, [MaybeNullWhen(false)] out KeyStageSubject parsed)
-    {
-        if (value == null)
-        {
-            parsed = default;
-            return false;
-        }
-
-        var re = new Regex(@"(KeyStage[\d])-(.*)").Match(value);
-
-        if (!re.Success)
-        {  // Subject must be of the form KS1-English
-            parsed = default;
-            return false;
-        }
-
-        if (!Enum.TryParse<KeyStage>(re.Groups[1].Value, out var ks))
-        { // Value is not a valid key stage
-            parsed = default;
-            return false;
-        }
-
-        parsed = new KeyStageSubject(ks, re.Groups[2].Value);
-        return true;
-    }
-}
-
-public enum ReferrerList
-{
-    SearchResults = 0,
-    FullList = 1
-}
-
-public enum KeyStage
-{
-    Unspecified = 0,
-
-    [Description("Key stage 1")]
-    KeyStage1 = 1,
-
-    [Description("Key stage 2")]
-    KeyStage2,
-
-    [Description("Key stage 3")]
-    KeyStage3,
-
-    [Description("Key stage 4")]
-    KeyStage4,
-}
-
-public enum Subject
-{
-    Unspecified = 0,
-
-    English,
-    Maths,
-    Science,
-
-    Humanities,
-    ModernForeignLanguages,
-}
-
-public enum TuitionType
-{
-    [Description("Any")]
-    Any = 0,
-
-    [Description("Online")]
-    Online = 1,
-
-    [Description("In School")]
-    InSchool,
-}
