+++ UI/Pages/TuitionPartner.cshtml.cs
@@ -6,6 +6,7 @@ using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
+using UI.Models;
 
 namespace UI.Pages;
 
@@ -87,7 +88,7 @@ public class TuitionPartner : PageModel
         string[] TuitionTypes, string[] Ratios, Dictionary<int, GroupPrice> Prices,
         string Website, string PhoneNumber, string EmailAddress, string[] Address, bool HasSenProvision, bool IsVatCharged,
         LocalAuthorityDistrictCoverage[] LocalAuthorityDistricts,
-        Dictionary<TuitionType, Dictionary<KeyStage, Dictionary<string, Dictionary<int, decimal>>>> AllPrices, string LegalStatus)
+        Dictionary<Enums.TuitionType, Dictionary<Enums.KeyStage, Dictionary<string, Dictionary<int, decimal>>>> AllPrices, string LegalStatus)
     {
         public bool HasPricingVariation => Prices.Any(x => x.Value.HasVariation);
     }
@@ -105,20 +106,20 @@ public class TuitionPartner : PageModel
 
     public class QueryHandler : IRequestHandler<Query, Command?>
     {
-        private readonly ILocationFilterService locationService;
-        private readonly INtpDbContext db;
-        private readonly ILogger<TuitionPartner> logger;
+        private readonly ILocationFilterService _locationService;
+        private readonly INtpDbContext _db;
+        private readonly ILogger<TuitionPartner> _logger;
 
         public QueryHandler(ILocationFilterService locationService, INtpDbContext db, ILogger<TuitionPartner> logger)
         {
-            this.locationService = locationService;
-            this.db = db;
-            this.logger = logger;
+            _locationService = locationService;
+            _db = db;
+            _logger = logger;
         }
 
         public async Task<Command?> Handle(Query request, CancellationToken cancellationToken)
         {
-            var queryable = this.db.TuitionPartners
+            var queryable = _db.TuitionPartners
                 .Include(e => e.SubjectCoverage)
                 .ThenInclude(e => e.Subject)
                 .Include(x => x.Prices)
@@ -140,19 +141,19 @@ public class TuitionPartner : PageModel
 
             if (tp == null) return null;
 
-            var subjects = tp.SubjectCoverage.Select(x => x.Subject).Distinct().GroupBy(x => x.KeyStageId).Select(x => $"{((KeyStage)x.Key).DisplayName()} - {x.DisplayList()}");
+            var subjects = tp.SubjectCoverage.Select(x => x.Subject).Distinct().GroupBy(x => x.KeyStageId).Select(x => $"{((Enums.KeyStage)x.Key).DisplayName()} - {x.DisplayList()}");
             string? ladCode = null;
             if (!string.IsNullOrEmpty(request.Postcode))
             {
-                var location = await locationService.GetLocationFilterParametersAsync(request.Postcode!);
+                var location = await _locationService.GetLocationFilterParametersAsync(request.Postcode!);
                 ladCode = location?.LocalAuthorityDistrictCode;
             }
             var types = await GetTuitionTypesCovered(ladCode, tp, cancellationToken);
-            if (types == null || types.Count() == 0)
+            if (types == null || !types.Any())
             {
                 //If no data returned then postcode is invalid, has been changed so does not apply for the TP
                 //or issue calling GetLocationFilterParametersAsync (calling postcode.io)
-                logger.LogWarning("Issue getting TuitionTypesCovered (invalid postcode, changed postcode or issue calling postcode.io) for postcode '{Postcode}' and TP '{Name}'", request.Postcode, tp.Name);
+                _logger.LogWarning("Issue getting TuitionTypesCovered (invalid postcode, changed postcode or issue calling postcode.io) for postcode '{Postcode}' and TP '{Name}'", request.Postcode, tp.Name);
                 types = await GetTuitionTypesCovered(null, tp, cancellationToken);
             }
             var ratios = tp.Prices.Select(x => x.GroupSize).Distinct().Select(x => $"1 to {x}");
@@ -185,7 +186,7 @@ public class TuitionPartner : PageModel
             Domain.TuitionPartner tp,
             CancellationToken cancellationToken)
         {
-            var coverageQuery = this.db
+            var coverageQuery = _db
                 .LocalAuthorityDistrictCoverage
                 .Where(e => e.TuitionPartnerId == tp.Id);
 
@@ -207,14 +208,14 @@ public class TuitionPartner : PageModel
         {
             if (!request.ShowLocationsCovered) return Array.Empty<LocalAuthorityDistrictCoverage>();
 
-            var coverage = await this.db.LocalAuthorityDistrictCoverage.Where(e => e.TuitionPartnerId == tpId)
+            var coverage = await _db.LocalAuthorityDistrictCoverage.Where(e => e.TuitionPartnerId == tpId)
                 .ToArrayAsync();
 
             var coverageDictionary = coverage
                 .GroupBy(e => e.TuitionTypeId)
-                .ToDictionary(e => (TuitionType)e.Key, e => e.ToDictionary(x => x.LocalAuthorityDistrictId, x => x));
+                .ToDictionary(e => (Enums.TuitionType)e.Key, e => e.ToDictionary(x => x.LocalAuthorityDistrictId, x => x));
 
-            var regions = await this.db.Regions
+            var regions = await _db.Regions
                 .Include(e => e.LocalAuthorityDistricts.OrderBy(x => x.Code))
                 .OrderBy(e => e.Name)
                 .ToDictionaryAsync(e => e, e => e.LocalAuthorityDistricts);
@@ -225,8 +226,8 @@ public class TuitionPartner : PageModel
             {
                 foreach (var lad in lads)
                 {
-                    var inSchool = coverageDictionary.ContainsKey(TuitionType.InSchool) && coverageDictionary[TuitionType.InSchool].ContainsKey(lad.Id);
-                    var online = coverageDictionary.ContainsKey(TuitionType.Online) && coverageDictionary[TuitionType.Online].ContainsKey(lad.Id);
+                    var inSchool = coverageDictionary.ContainsKey(Enums.TuitionType.InSchool) && coverageDictionary[Enums.TuitionType.InSchool].ContainsKey(lad.Id);
+                    var online = coverageDictionary.ContainsKey(Enums.TuitionType.Online) && coverageDictionary[Enums.TuitionType.Online].ContainsKey(lad.Id);
                     result[lad.Id] = new LocalAuthorityDistrictCoverage(region.Name, lad.Code, lad.Name, inSchool, online);
                 }
             }
@@ -276,20 +277,20 @@ public class TuitionPartner : PageModel
             }
         }
 
-        private async Task<Dictionary<TuitionType, Dictionary<KeyStage, Dictionary<string, Dictionary<int, decimal>>>>> GetFullPricing(Query request, ICollection<Price> prices)
+        private async Task<Dictionary<Enums.TuitionType, Dictionary<Enums.KeyStage, Dictionary<string, Dictionary<int, decimal>>>>> GetFullPricing(Query request, ICollection<Price> prices)
         {
             if (!request.ShowFullPricing) return new();
 
-            var fullPricing = new Dictionary<TuitionType, Dictionary<KeyStage, Dictionary<string, Dictionary<int, decimal>>>>();
+            var fullPricing = new Dictionary<Enums.TuitionType, Dictionary<Enums.KeyStage, Dictionary<string, Dictionary<int, decimal>>>>();
 
-            foreach (var tuitionType in new[] { TuitionType.InSchool, TuitionType.Online })
+            foreach (var tuitionType in new[] { Enums.TuitionType.InSchool, Enums.TuitionType.Online })
             {
-                fullPricing[tuitionType] = new Dictionary<KeyStage, Dictionary<string, Dictionary<int, decimal>>>();
-                foreach (var keyStage in new[] { KeyStage.KeyStage1, KeyStage.KeyStage2, KeyStage.KeyStage3, KeyStage.KeyStage4 })
+                fullPricing[tuitionType] = new Dictionary<Enums.KeyStage, Dictionary<string, Dictionary<int, decimal>>>();
+                foreach (var keyStage in new[] { Enums.KeyStage.KeyStage1, Enums.KeyStage.KeyStage2, Enums.KeyStage.KeyStage3, Enums.KeyStage.KeyStage4 })
                 {
                     fullPricing[tuitionType][keyStage] = new Dictionary<string, Dictionary<int, decimal>>();
 
-                    var keyStageSubjects = await this.db.Subjects.Where(e => e.KeyStageId == (int)keyStage).OrderBy(e => e.Name).ToArrayAsync();
+                    var keyStageSubjects = await _db.Subjects.Where(e => e.KeyStageId == (int)keyStage).OrderBy(e => e.Name).ToArrayAsync();
                     foreach (var subject in keyStageSubjects)
                     {
                         fullPricing[tuitionType][keyStage][subject.Name] = new Dictionary<int, decimal>();
@@ -299,8 +300,8 @@ public class TuitionPartner : PageModel
 
             foreach (var price in prices)
             {
-                var tuitionType = (TuitionType)price.TuitionTypeId;
-                var keyStage = (KeyStage)price.Subject.KeyStageId;
+                var tuitionType = (Enums.TuitionType)price.TuitionTypeId;
+                var keyStage = (Enums.KeyStage)price.Subject.KeyStageId;
                 var subjectName = price.Subject.Name;
 
                 fullPricing[tuitionType][keyStage][subjectName][price.GroupSize] = price.HourlyRate;
