+++ UI/Pages/TuitionPartnerLogo.cshtml.cs
@@ -8,13 +8,13 @@ namespace UI.Pages;
 [ResponseCache(Duration = 60 * 60 * 24 * 7, Location = ResponseCacheLocation.Any)]
 public class TuitionPartnerLogo : PageModel
 {
-    private readonly INtpDbContext ntpDb;
+    private readonly INtpDbContext _db;
 
-    public TuitionPartnerLogo(INtpDbContext ntpDb) => this.ntpDb = ntpDb;
+    public TuitionPartnerLogo(INtpDbContext db) => _db = db;
 
     public async Task<IActionResult> OnGet(string id)
     {
-        var logo = await ntpDb.TuitionPartners
+        var logo = await _db.TuitionPartners
             .Include(x => x.Logo)
             .Where(x => x.SeoUrl == id)
             .Select(x => x.Logo)
